MNT: move metadataserver into tree and get tests running
