--- connection.yml
@@ -1,5 +1,0 @@
-# metadataservice configuration
-# DO NO EDIT unless authorized
-host: localhost
-port: 7778
-timezone: US/Eastern
