--- metadataservice/server/__init__.py
@@ -1,6 +1,0 @@
-__author__ = 'arkilic'
-from metadataservice._version import get_versions
-
-
-__version__ = get_versions()['version']
-del get_versions
